Atomic configurations

We define a configuration to be a set of orbitals with their associated occupation (i.e. the number of electron on that orbital). We can represent a particular configuration with an instance of the Configuration type.

AtomicLevels.Configuration — Type.

struct Configuration{<:AbstractOrbital}

Represents a configuration – a set of orbitals and their associated occupation number. Furthermore, each orbital can be in one of the following states: :open, :closed or :inactive.

Constructors

Configuration(orbitals :: Vector{<:AbstractOrbital}, occupancy :: Vector{Int}, states :: Vector{Symbol})
Configuration(orbitals :: Vector{Tuple{<:AbstractOrbital, Int, Symbol}})

In the first case, the paramaters of each orbital have to be passed as separate vectors, and the orbitals and occupancy have to be of the same length. The states vector can be shorter and then the latter orbitals that were not explicitly specified by states are assumed to be :open.

The second constructor allows you to pass a vector of tuples instead, where each tuple is a triplet (orbital :: AbstractOrbital, occupancy :: Int, state :: Symbol) corresponding to each orbital.

In all cases, all the orbitals have to be distinct. The orbitals in the configuration will be sorted according to the ordering defined for the particular AbstractOrbital.

The @c_str and @rc_str string macros can be used to conveniently construct configurations:

AtomicLevels.@c_str — Macro.

@c_str -> Configuration{Orbital}

Construct a Configuration, representing a non-relativistic configuration, out of a string.

julia> c"1s2 2s"
1s² 2s

julia> c"[Kr] 4d10 5s2 4f2"
[Kr]ᶜ 4d¹⁰ 4f² 5s²

AtomicLevels.@rc_str — Macro.

@rc_str -> Configuration{RelativisticOrbital}

Construct a Configuration representing a relativistic configuration out of a string.

julia> rc"[Ne] 3s 3p- 3p"
[Ne]ᶜ 3s 3p⁻ 3p

julia> rc"[Ne] 3s 3p-2 3p4"
[Ne]ᶜ 3s 3p⁻² 3p⁴

Interface

For example, it is possible to index into a configuration, including with a range of indices, returning a sub-configuration consisting of only those orbitals. With an integer index, an (orbital, occupancy, state) tuple is returned.

julia> config = c"1s2c 2si 2p3"
[He]ᶜ 2sⁱ 2p³

julia> config[2]
(2s, 1, :inactive)

julia> config[1:2]
[He]ᶜ 2sⁱ

julia> config[[3,1]]
[He]ᶜ 2p³

The configuration can also be iterated over. Each item is a (orbital, occupancy, state) tuple.

julia> for (o, nelec, s) in config
           @show o, nelec, s
       end
(o, nelec, s) = (1s, 2, :closed)
(o, nelec, s) = (2s, 1, :inactive)
(o, nelec, s) = (2p, 3, :open)

Various other methods exist to manipulate or transform configurations or to query them for information.

AtomicLevels.num_electrons — Method.

num_electrons(c::Configuration) -> Int

Return the number of electrons in the configuration.

julia> num_electrons(c"1s2")
2

julia> num_electrons(rc"[Kr] 5s2 5p-2 5p2")
42

AtomicLevels.num_electrons — Method.

num_electrons(c::Configuration, o::AbstractOrbital) -> Int

Returns the number of electrons on orbital o in configuration c. If o is not part of the configuration, returns 0.

julia> num_electrons(c"1s 2s2", o"2s")
2

julia> num_electrons(rc"[Rn] Qf-5 Pf3", ro"Qf-")
5

julia> num_electrons(c"[Ne]", o"3s")
0

Base.delete! — Function.

delete!(c::Configuration, o::AbstractOrbital)

Remove the entire subshell corresponding to orbital o from configuration c.

julia> delete!(c"[Ar] 4s2 3d10 4p2", o"4s")
[Ar]ᶜ 3d¹⁰ 4p²

Base.:+ — Function.

+(::Configuration, ::Configuration)

Add two configurations together. If both configuration have an orbital, the number of electrons gets added together, but in this case the status of the orbitals must match.

julia> c"1s" + c"2s"
1s 2s

julia> c"1s" + c"1s"
1s²

Base.:- — Function.

-(configuration::Configuration, orbital::AbstractOrbital[, n=1])

Remove n electrons in the orbital orbital from the configuration configuration. If the orbital had previously been :closed or :inactive, it will now be :open.

Base.close — Function.

close(c::Configuration)

Return a corresponding configuration where where all the orbitals are marked :closed.

See also: close!

AtomicLevels.close! — Function.

close!(c::Configuration)

Marks all the orbitals in configuration c to be closed.

See also: close

Base.fill — Function.

fill(c::Configuration)

Returns a corresponding configuration where the orbitals are completely filled (as determined by degeneracy).

See also: fill!

Base.fill! — Function.

fill!(c::Configuration)

Sets all the occupancies in configuration c to maximum, as determined by the degeneracy function.

See also: fill

Base.in — Function.

in(o::AbstractOrbital, c::Configuration) -> Bool

Checks if orbital o is part of configuration c.

julia> in(o"2s", c"1s2 2s2")
true

julia> o"2p" ∈ c"1s2 2s2"
false

Base.filter — Function.

filter(f, c::Configuration) -> Configuration

Filter out the orbitals from configuration c for which the predicate f returns false. The predicate f needs to take three arguments: orbital, occupancy and state.

julia> filter((o,occ,s) -> o.ℓ == 1, c"[Kr]")
2p⁶ᶜ 3p⁶ᶜ 4p⁶ᶜ

Base.count — Function.

count(::Configuration) -> Int

Return the number of electrons in the configuration.

julia> count(c"[Kr] 5s")
37

AtomicLevels.core — Function.

core(::Configuration) -> Configuration

Return the core configuration (i.e. the sub-configuration of all the orbitals that are marked :closed).

julia> core(c"1s2c 2s2c 2p6c 3s2")
[Ne]ᶜ

julia> core(c"1s2 2s2")
∅

julia> core(c"1s2 2s2c 2p6c")
2s²ᶜ 2p⁶ᶜ

AtomicLevels.peel — Function.

peel(::Configuration) -> Configuration

Return the non-core part of the configuration (i.e. orbitals not marked :closed).

julia> peel(c"1s2c 2s2c 2p3")
2p³

julia> peel(c"[Ne] 3s 3p3")
3s 3p³

AtomicLevels.active — Function.

active(::Configuration) -> Configuration

Return the part of the configuration marked :open.

julia> active(c"1s2c 2s2i 2p3i 3s2")
3s²

AtomicLevels.inactive — Function.

inactive(::Configuration) -> Configuration

Return the part of the configuration marked :inactive.

julia> inactive(c"1s2c 2s2i 2p3i 3s2")
2s²ⁱ 2p³ⁱ

AtomicLevels.bound — Function.

bound(::Configuration) -> Configuration

Return the bound part of the configuration (see also isbound).

julia> bound(c"1s2 2s2 2p4 Ks2 Kp1")
1s² 2s² 2p⁴

AtomicLevels.continuum — Function.

continuum(::Configuration) -> Configuration

Return the non-bound (continuum) part of the configuration (see also isbound).

julia> continuum(c"1s2 2s2 2p4 Ks2 Kp1")
Ks² Kp

AtomicLevels.parity — Method.

parity(::Configuration) -> Parity

Return the parity of the configuration.

julia> parity(c"1s 2p")
odd

julia> parity(c"1s 2p2")
even

See also: Parity

AtomicLevels.nonrelconfiguration — Function.

nonrelconfiguration(c::Configuration{<:RelativisticOrbital}) -> Configuration{<:Orbital}

Reduces a relativistic configuration down to the corresponding non-relativistic configuration.

julia> c = rc"1s2 2p-2 2s 2p2 3s2 3p-"
1s² 2s 2p⁻² 2p² 3s² 3p⁻

julia> nonrelconfiguration(c)
1s² 2s 2p⁴ 3s² 3p

Generating configuration lists

The ⊗ operator can be used to easily generate lists of configurations from existing pieces. E.g. to create all the valence configurations on top of an closed core, you only need to write

julia> c"[Ne]" ⊗ [c"3s2", c"3s 3p", c"3p2"]
3-element Array{Configuration{Orbital{Int64}},1}:
 [Ne]ᶜ 3s²
 [Ne]ᶜ 3s 3p
 [Ne]ᶜ 3p²

That can be combined with the @rcs_str string macro to easily generate all possible relativistic configurations from a non-relativistic definition:

julia> rc"[Ne] 3s2" ⊗ rcs"3p2"
3-element Array{Configuration{RelativisticOrbital{Int64}},1}:
 [Ne]ᶜ 3s² 3p⁻²
 [Ne]ᶜ 3s² 3p⁻ 3p
 [Ne]ᶜ 3s² 3p²

AtomicLevels.:⊗ — Function.

⊗(::Union{Configuration, Vector{Configuration}}, ::Union{Configuration, Vector{Configuration}})

Given two collections of Configurations, it creates an array of Configurations with all possible juxtapositions of configurations from each collection.

Examples

julia> c"1s" ⊗ [c"2s2", c"2s 2p"]
2-element Array{Configuration{Orbital{Int64}},1}:
 1s 2s²
 1s 2s 2p

julia> [rc"1s", rc"2s"] ⊗ [rc"2p-", rc"2p"]
4-element Array{Configuration{RelativisticOrbital{Int64}},1}:
 1s 2p⁻
 1s 2p
 2s 2p⁻
 2s 2p

AtomicLevels.@rcs_str — Macro.

@rcs_str -> Vector{Configuration{RelativisticOrbital}}

Construct a Vector of all Configurations corresponding to the non-relativistic nℓ orbital with the given occupancy from the input string.

The string is assumed to have the following syntax: $(n)$(ℓ)$(occupancy), where n and occupancy are integers, and ℓ is in spectroscopic notation.

Examples

julia> rcs"3p2"
3-element Array{Configuration{RelativisticOrbital{N}} where N,1}:
 3p⁻²
 3p⁻ 3p
 3p²

Spin configurations

AtomicLevels.spin_configurations — Function.

spin_configurations(configuration)

Generate all possible configurations of spin-orbitals from configuration, i.e. all permissible values for the quantum numbers n, ℓ, mℓ, ms for each electron. Example:

julia> spin_configurations(c"1s2")
1-element Array{Configuration{SpinOrbital{Orbital{Int64}}},1}:
 1s²

julia> spin_configurations(c"1s ks")
4-element Array{Configuration{SpinOrbital},1}:
 1s₀α ks₀α
 1s₀β ks₀α
 1s₀α ks₀β
 1s₀β ks₀β

spin_configurations(configurations)

For each configuration in configurations, generate all possible configurations of spin-orbitals.

AtomicLevels.substitutions — Function.

substitutions(src::Configuration{<:SpinOrbital}, dst::Configuration{<:SpinOrbital})

Find all orbital substitutions going from spin-configuration src to configuration dst.