Atomic orbitals

Atomic orbitals, i.e. single particle states with well-defined orbital or total angular momentum, are usually the basic building blocks of atomic states. AtomicLevels provides various types and methods to conveniently label the orbitals.

Orbital types

AtomicLevels provides two basic types for labelling atomic orbitals: Orbital and RelativisticOrbital. Stricly speaking, these types do not label orbitals, but groups of orbitals with the same angular symmetry and radial behaviour (i.e. a subshell).

All orbitals are subtypes of AbstractOrbital. Types and methods that work on generic orbitals can dispatch on that.

AtomicLevels.Orbital — Type.

struct Orbital{N <: AtomicLevels.MQ} <: AbstractOrbital

Label for an atomic orbital with a principal quantum number n::N and orbital angular momentum ℓ.

The type parameter N has to be such that it can represent a proper principal quantum number (i.e. a subtype of AtomicLevels.MQ).

Properties

The following properties are part of the public API:

.n :: N – principal quantum number $n$

.ℓ :: Int – the orbital angular momentum $\ell$

Constructors

Orbital(n::Int, ℓ::Int)
Orbital(n::Symbol, ℓ::Int)

Construct an orbital label with principal quantum number n and orbital angular momentum ℓ. If the principal quantum number n is an integer, it has to positive and the angular momentum must satisfy 0 <= ℓ < n.

julia> Orbital(1, 0)
1s

julia> Orbital(:K, 2)
Kd

AtomicLevels.RelativisticOrbital — Type.

struct RelativisticOrbital{N <: AtomicLevels.MQ} <: AbstractOrbital

Label for an atomic orbital with a principal quantum number n::N and well-defined total angular momentum $j$. The angular component of the orbital is labelled by the $(\ell, j)$ pair, conventionally written as $\ell_j$ (e.g. $p_{3/2}$).

The $\ell$ and $j$ can not be arbitrary, but must satisfy $j = \ell \pm 1/2$. Internally, the $\kappa$ quantum number, which is a unique integer corresponding to every physical $(\ell, j)$ pair, is used to label each allowed pair. When $j = \ell \pm 1/2$, the corresponding $\kappa = \mp(j + 1/2)$.

When printing and parsing RelativisticOrbitals, the notation nℓ and nℓ- is used (e.g. 2p and 2p-), corresponding to the orbitals with $j = \ell + 1/2$ and $j = \ell - 1/2$, respectively.

The type parameter N has to be such that it can represent a proper principal quantum number (i.e. a subtype of AtomicLevels.MQ).

Properties

The following properties are part of the public API:

.n :: N – principal quantum number $n$

.κ :: Int – $\kappa$ quantum number

.ℓ :: Int – the orbital angular momentum label $\ell$

.j :: HalfInteger – total angular momentum $j$

julia> orb = ro"5g-"
5g⁻

julia> orb.n
5

julia> orb.j
7/2

julia> orb.ℓ
4

Constructors

RelativisticOrbital(n::Integer, κ::Integer)
RelativisticOrbital(n::Symbol, κ::Integer)
RelativisticOrbital(n, ℓ::Integer, j::Real)

Construct an orbital label with the quantum numbers n and κ. If the principal quantum number n is an integer, it has to positive and the orbital angular momentum must satisfy 0 <= ℓ < n. Instead of κ, valid ℓ and j values can also be specified instead.

julia> RelativisticOrbital(1, 0, 1//2)
1s

julia> RelativisticOrbital(2, -1)
2s

julia> RelativisticOrbital(:K, 2, 3//2)
Kd⁻

AtomicLevels.AbstractOrbital — Type.

abstract type AbstractOrbital

Abstract supertype of all orbital types.

The SpinOrbital type can be used to fully qualify all the quantum numbers (that is, also $m_\ell$ and $m_s$) of an Orbital. It represent a since, distinct orbital.

AtomicLevels.SpinOrbital — Type.

struct SpinOrbital{O<:Orbital} <: AbstractOrbital

Spin orbitals are fully characterized orbitals, i.e. the quantum numbers n, ℓ, mℓ and ms are all specified.

The string macros @o_str and @ro_str can be used to conveniently contruct orbitals, while @os_str and @ros_str can be used to construct whole lists of them very easily.

AtomicLevels.@o_str — Macro.

@o_str -> Orbital

A string macro to construct an Orbital from the canonical string representation.

julia> o"1s"
1s

julia> o"Fd"
Fd

AtomicLevels.@ro_str — Macro.

@ro_str -> RelativisticOrbital

A string macro to construct an RelativisticOrbital from the canonical string representation.

julia> ro"1s"
1s

julia> ro"2p-"
2p⁻

julia> ro"Kf-"
Kf⁻

AtomicLevels.@os_str — Macro.

@os_str -> Vector{Orbital}

Can be used to easily construct a list of Orbitals.

julia> os"5[d] 6[s-p] k[7-10]"
7-element Array{Orbital,1}:
 5d
 6s
 6p
 kk
 kl
 km
 kn

AtomicLevels.@ros_str — Macro.

@ros_str -> Vector{RelativisticOrbital}

Can be used to easily construct a list of RelativisticOrbitals.

julia> ros"2[s-p] 3[p] k[0-d]"
10-element Array{RelativisticOrbital,1}:
 2s
 2p⁻
 2p
 3p⁻
 3p
 ks
 kp⁻
 kp
 kd⁻
 kd

Methods

Base.isless — Function.

isless(a::Orbital, b::Orbital)

Compares the orbitals a and b to decide which one comes before the other in a configuration.

Examples

julia> o"1s" < o"2s"
true

julia> o"1s" < o"2p"
true

julia> o"ks" < o"2p"
false

AtomicLevels.degeneracy — Function.

degeneracy(orbital::Orbital)

Returns the degeneracy of orbital which is 2(2ℓ+1)

Examples

julia> degeneracy(o"1s")
2

julia> degeneracy(o"2p")
6

AtomicLevels.parity — Method.

parity(orbital::Orbital)

Returns the parity of orbital, defined as (-1)^ℓ.

Examples

julia> parity(o"1s")
even

julia> parity(o"2p")
odd

AtomicLevels.symmetry — Function.

symmetry(orbital::Orbital)

Returns the symmetry for orbital which is simply ℓ.

AtomicLevels.isbound — Function.

isbound(::Orbital)

Returns true is the main quantum number is an integer, false otherwise.

julia> isbound(o"1s")
true

julia> isbound(o"ks")
false

AtomicLevels.mℓrange — Function.

mℓrange(orbital::Orbital)

Returns the range of valid values of mℓ for orbital.

Examples

julia> mℓrange(o"2p")
-1:1

AtomicLevels.spin_orbitals — Function.

spin_orbitals(orbital)

Generate all permissible spin-orbitals for a given orbital, e.g. 2p -> 2p ⊗ mℓ = {-1,0,1} ⊗ s = {α,β}

Examples

julia> spin_orbitals(o"2p")
6-element Array{SpinOrbital{Orbital{Int64}},1}:
 2p₋₁α
 2p₋₁β
 2p₀α
 2p₀β
 2p₁α
 2p₁β

AtomicLevels.@κ_str — Macro.

@κ_str -> Int

A string macro to convert the canonical string representation of a $\ell_j$ angular label (i.e. ℓ- or ℓ) into the corresponding $\kappa$ quantum number.

julia> κ"s", κ"p-", κ"p"
(-1, 1, -2)