Term symbols

AtomicLevels provides types and methods to work and determine term symbols. The "Term symbol" and "Angular momentum coupling" Wikipedia articles give a good basic overview of the terminology.

For term symbols in LS coupling, AtomicLevels provides the Term type.

AtomicLevels.Term — Type.

struct Term

Represent a term symbol ${}^{2S+1}L_{J}$ with specific parity in LS-coupling. As determining valid $J$ values is simple for given $S$ and $L$ ($|L - S| \leq J \leq L+S$), it is not specified.

Constructors

Term(L::Real, S::Real, parity::Union{Parity,Integer})

Constructs a Term object with the given $L$ and $S$ quantum numbers and parity. L and S both have to be convertible to HalfIntegers and parity must be of type Parity or ±1.

The Term objects can also be constructed with the @T_str string macro.

AtomicLevels.@T_str — Macro.

@T_str -> Term

Constructs a Term object out of its canonical string representation.

julia> T"1S"
¹S

julia> T"4Po"
⁴Pᵒ

julia> T"2[3/2]o" # jK coupling, common in noble gases
²[3/2]ᵒ

The terms function can be used to generate all possible term symbols. In the case of relativistic orbitals, the term symbols are simply the valid $J$ values, represented with the HalfInteger type.

AtomicLevels.terms — Function.

terms(orb::Orbital, w::Int=one(Int))

Returns a list of valid LS term symbols for the orbital orb with w occupancy.

Examples

julia> terms(o"3d", 3)
8-element Array{Term,1}:
 ²P
 ²D
 ²D
 ²F
 ²G
 ²H
 ⁴P
 ⁴F

terms(config)

Generate all final $LS$ terms for config.

Examples

julia> terms(c"1s")
1-element Array{Term,1}:
 ²S

julia> terms(c"1s 2p")
2-element Array{Term,1}:
 ¹Pᵒ
 ³Pᵒ

julia> terms(c"[Ne] 3d3")
7-element Array{Term,1}:
 ²P
 ²D
 ²F
 ²G
 ²H
 ⁴P
 ⁴F

terms(o::RelativisticOrbital, w = 1) -> Vector{HalfInt}

Returns a sorted list of valid $J$ values of w equivalent $jj$-coupled particles on orbital o (i.e. oʷ).

When there are degeneracies (i.e. multiple states with the same $J$ and $M$ quantum numbers), the corresponding $J$ value is repeated in the output array.

Examples

julia> terms(ro"3d", 3)
3-element Array{HalfIntegers.Half{Int64},1}:
 3/2
 5/2
 9/2

julia> terms(ro"3d-", 3)
1-element Array{HalfIntegers.Half{Int64},1}:
 3/2

julia> terms(ro"4f", 4)
8-element Array{HalfIntegers.Half{Int64},1}:
 0
 2
 2
 4
 4
 5
 6
 8

Term multiplicity and intermediate terms

For subshells starting with d³, the possible terms may occur more than once (multiplicity higher than one), corresponding to different physical states. These arise from different sequences of coupling the $w$ equivalent electrons of the same $\ell$, and are distinguished using a seniority number, which the IntermediateTerm type implements. For partially filled f shells, seniority is not enough to distinguish all possible couplings. Using count_terms, we can see that e.g. the ²Dᵒ has higher multiplicity than can be described with seniority only:

julia> count_terms(o"4f", 3, T"2Do")
2

julia> count_terms(o"4f", 5, T"2Do")
5

AtomicLevels.IntermediateTerm — Type.

IntermediateTerm(term, seniority)

Represents a term together with its seniority quantum number.

AtomicLevels.intermediate_terms — Function.

intermediate_terms(orb::Orbital, w::Int=one(Int))

Generates all IntermediateTerm for a given non-relativstic orbital orb and occupation w.

Examples

julia> intermediate_terms(o"2p", 2)
3-element Array{IntermediateTerm,1}:
 ₀¹S
 ₂¹D
 ₂³P

The preceding subscript is the seniority number, which indicates at which occupancy a certain term is first seen, cf.

julia> intermediate_terms(o"3d", 1)
1-element Array{IntermediateTerm,1}:
 ₁²D

julia> intermediate_terms(o"3d", 3)
8-element Array{IntermediateTerm,1}:
 ₁²D
 ₃²P
 ₃²D
 ₃²F
 ₃²G
 ₃²H
 ₃⁴P
 ₃⁴F

In the second case, we see both ₁²D and ₃²D, since there are two ways of coupling 3 d electrons to a ²D symmetry.

intermediate_terms(config)

Generate the intermediate terms for each subshell of config.

Examples

julia> intermediate_terms(c"1s 2p3")
2-element Array{Array{IntermediateTerm,1},1}:
 [₁²S]
 [₁²Pᵒ, ₃²Dᵒ, ₃⁴Sᵒ]

julia> intermediate_terms(rc"3d2 5g3")
2-element Array{Array{HalfIntegers.Half{Int64},1},1}:
 [0, 2, 4]
 [3/2, 5/2, 7/2, 9/2, 9/2, 11/2, 13/2, 15/2, 17/2, 21/2]

AtomicLevels.count_terms — Function.

count_terms(orb, occ, term)

Count how many times term occurs among the valid terms of orb^occ. For example:

julia> count_terms(o"1s", 2, T"1S")
1

Internal implementation of term multiplicity calculation

AtomicLevels.jl uses the algorithm presented in

Alternative mathematical technique to determine LS spectral terms by Xu Renjun and Dai Zhenwen, published in JPhysB, 2006.

doi:10.1088/0953-4075/39/16/007

to compute the multiplicity of individual subshells, beyond the trivial cases of a single electron or a filled subshell. These routines need not be used directly, instead use terms and count_terms.

In the following, $S'=2S\in\mathbb{Z}$ and $M_S'=2M_S\in\mathbb{Z}$, as in the original article.

AtomicLevels.xu_terms — Function.

xu_terms(ℓ, w, p)

Return all term symbols for the orbital ℓʷ and parity p; the term multiplicity is computed using AtomicLevels.Xu.X.

Examples

julia> AtomicLevels.xu_terms(3, 3, parity(c"3d3"))
17-element Array{Term,1}:
 ²P
 ²D
 ²D
 ²F
 ²F
 ²G
 ²G
 ²H
 ²H
 ²I
 ²K
 ²L
 ⁴S
 ⁴D
 ⁴F
 ⁴G
 ⁴I

AtomicLevels.Xu.X — Function.

X(N, ℓ, S′, L)

Calculate the multiplicity of the term $^{2S+1}L$ ($S'=2S$) for the orbital ℓ with occupancy N, according to the formula:

\[\begin{aligned} X(N, \ell, S', L) =&+ A(N, \ell,\ell,S',L)\\ &- A(N, \ell,\ell,S',L+1)\\ &+A(N, \ell,\ell,S'+2,L+1)\\ &- A(N, \ell,\ell,S'+2,L) \end{aligned}\]

Note that this is not correct for filled (empty) shells, for which the only possible term trivially is ¹S.

Examples

julia> AtomicLevels.Xu.X(1, 0, 1, 0) # Multiplicity of ²S term for s¹
1

julia> AtomicLevels.Xu.X(3, 3, 1, 3) # Multiplicity of ²D term for d³
2

AtomicLevels.Xu.A — Function.

$A(N,\ell,\ell_b,M_S',M_L)$ obeys four different cases:

Case 1

$M_S'=1$, $|M_L|\leq\ell$, and $N=1$:

\[A(1,\ell,\ell_b,1,M_L) = 1\]

Case 2

$\{M_S'\}={2-N,4-N,...,N-2}$, $|M_L| \leq f\left(\frac{N-M_S'}{2}-1\right)+f\left(\frac{N+M_S'}{2}-1\right)$, and $1<N\leq 2\ell+1$:

\[\begin{aligned} A(N,\ell,\ell,M_S',M_L) = \sum_{M_{L-}\max\left\{-f\left(\frac{N-M_S'}{2}-1\right),M_L-f\left(\frac{N+M_S'}{2}-1\right)\right\}} ^{\min\left\{f\left(\frac{N-M_S'}{2}-1\right),M_L+f\left(\frac{N+M_S'}{2}-1\right)\right\}} \Bigg\{A\left(\frac{N-M_S'}{2},\ell,\ell,\frac{N-M_S'}{2},M_{L-}\right)\\ \times A\left(\frac{N+M_S'}{2},\ell,\ell,\frac{N+M_S'}{2},M_L-M_{L-}\right)\Bigg\} \end{aligned}\]

Case 3

$M_S'=N$, $|M_L|\leq f(N-1)$, and $1<N\leq 2\ell+1$:

\[A(N,\ell,\ell_b,N,M_L) = \sum_{M_{L_I} = \left\lfloor{\frac{M_L-1}{N}+\frac{N+1}{2}}\right\rfloor} ^{\min\{\ell_b,M_L+f(N-2)\}} A(N-1,\ell,M_{L_I}-1,N-1,M_L-M_{L_I})\]

Case 4

else:

\[A(N,\ell,\ell_b,M_S',M_L) = 0\]

AtomicLevels.Xu.f — Function.

f(n,ℓ)

\[f(n,\ell)=\begin{cases} \displaystyle\sum_{m=0}^n \ell-m, & n\geq0\\ 0, & n<0 \end{cases}\]

Term couplings

The angular momentum coupling method is based on the vector model, where two angular momenta can be combined via vector addition to form a total angular momentum:

\[\vec{J} = \vec{L} + \vec{S},\]

where the length of the resultant momentum $\vec{J}$ obeys

\[|L-S| \leq J \leq L+S.\]

Relations such as these are used to couple the term symbols in both $LS$ and $jj$ coupling; however, not all values of $J$ predicted by the vector model are valid physical states, see couple_terms.

To generate the possible terms of a configuration, all the possible terms of the individual subshells, have to be coupled together to form the final terms; this is done from left-to-right. When generating all possible CSFs from a configuration, it is also necessary to find the intermediate couplings of the individual subshells. As an example, if we want to find the possible terms of 3p² 4s 5p², we first find the possible terms of the individual subshells:

julia> its = intermediate_terms(c"3p2 4s 5p2")
3-element Array{Array{IntermediateTerm,1},1}:
 [₀¹S, ₂¹D, ₂³P]
 [₁²S]
 [₀¹S, ₂¹D, ₂³P]

where the seniority numbers are indicated as preceding subscripts. We then need to couple each intermediate term of the first subshell with each of the second subshell, and couple each of the resulting terms with each of the third subshell, and so on. E.g. coupling the ₂³P intermediate term with ₁²S produces two terms:

julia> couple_terms(T"3P", T"2S")
2-element Array{Term,1}:
 ²P
 ⁴P

each of which need to be coupled with e.g. ₂¹D:

julia> couple_terms(T"2P", T"1D")
3-element Array{Term,1}:
 ²P
 ²D
 ²F

julia> couple_terms(T"4P", T"1D")
3-element Array{Term,1}:
 ⁴P
 ⁴D
 ⁴F

terms uses couple_terms (through AtomicLevels.final_terms) to produce all possible terms coupling trees, folding from left-to-right:

julia> a = couple_terms([T"1S", T"1D", T"3P"], [T"2S"])
4-element Array{Term,1}:
 ²S
 ²P
 ²D
 ⁴P

julia> couple_terms(a, [T"1S", T"1D", T"3P"])
12-element Array{Term,1}:
 ²S
 ²P
 ²D
 ²F
 ²G
 ⁴S
 ⁴P
 ⁴D
 ⁴F
 ⁶S
 ⁶P
 ⁶D

which gives the same result as

julia> terms(c"3p2 4s 5p2")
12-element Array{Term,1}:
 ²S
 ²P
 ²D
 ²F
 ²G
 ⁴S
 ⁴P
 ⁴D
 ⁴F
 ⁶S
 ⁶P
 ⁶D

Note that for the generation of final terms, the intermediate terms need not be kept (and their seniority is not important). However, for the generation of CSFs, we need to form all possible combinations of intermediate terms for each subshell, and couple them, again left-to-right, to form all possible coupling chains (each one corresponding to a unique physical state). E.g. for the last term of each subshell of 3p² 4s 5p²

julia> last.(its)
3-element Array{IntermediateTerm,1}:
 ₂³P
 ₁²S
 ₂³P

we find the following chains:

julia> intermediate_couplings(last.(its))
15-element Array{Array{Term,1},1}:
 [¹S, ³P, ²P, ²S]
 [¹S, ³P, ²P, ²P]
 [¹S, ³P, ²P, ²D]
 [¹S, ³P, ²P, ⁴S]
 [¹S, ³P, ²P, ⁴P]
 [¹S, ³P, ²P, ⁴D]
 [¹S, ³P, ⁴P, ²S]
 [¹S, ³P, ⁴P, ²P]
 [¹S, ³P, ⁴P, ²D]
 [¹S, ³P, ⁴P, ⁴S]
 [¹S, ³P, ⁴P, ⁴P]
 [¹S, ³P, ⁴P, ⁴D]
 [¹S, ³P, ⁴P, ⁶S]
 [¹S, ³P, ⁴P, ⁶P]
 [¹S, ³P, ⁴P, ⁶D]

AtomicLevels.couple_terms — Function.

couple_terms(t1, t2)

Generate all possible coupling terms between t1 and t2. It is assumed that t1 and t2 originate from non-equivalent electrons (i.e. from different subshells), since the vector model does not predict correct term couplings for equivalent electrons; some of the generated terms would violate the Pauli principle; cf. Cowan p. 108–109.

Examples

julia> couple_terms(T"1Po", T"2Se")
1-element Array{Term,1}:
 ²Pᵒ

julia> couple_terms(T"3Po", T"2Se")
2-element Array{Term,1}:
 ²Pᵒ
 ⁴Pᵒ

julia> couple_terms(T"3Po", T"2De")
6-element Array{Term,1}:
 ²Pᵒ
 ²Dᵒ
 ²Fᵒ
 ⁴Pᵒ
 ⁴Dᵒ
 ⁴Fᵒ

couple_terms(t1s, t2s)

Generate all coupling between all terms in t1s and all terms in t2s.

AtomicLevels.final_terms — Function.

final_terms(ts::Vector{<:Vector{<:Union{Term,Real}}})

Generate all possible final terms from the vector of vectors of individual subshell terms by coupling from left to right.

Examples

julia> ts = [[T"1S", T"3S"], [T"2P", T"2D"]]
2-element Array{Array{Term,1},1}:
 [¹S, ³S]
 [²P, ²D]

julia> AtomicLevels.final_terms(ts)
4-element Array{Term,1}:
 ²P
 ²D
 ⁴P
 ⁴D

AtomicLevels.intermediate_couplings — Function.

intermediate_couplings(its::Vector{IntermediateTerm,Integer,HalfInteger}, t₀ = T"1S")

Generate all intermediate coupling trees from the vector of intermediate terms its, starting from the initial term t₀.

Examples

julia> intermediate_couplings([IntermediateTerm(T"2S", 1), IntermediateTerm(T"2D", 1)])
2-element Array{Array{Term,1},1}:
 [¹S, ²S, ¹D]
 [¹S, ²S, ³D]

intermediate_couplings(J::Vector{<:Real}, j₀ = 0)

Examples

julia> intermediate_couplings([1//2, 3//2])
2-element Array{Array{HalfIntegers.Half{Int64},1},1}:
 [0, 1/2, 1]
 [0, 1/2, 2]