Other utilities

Parity

AtomicLevels defines the Parity type, which is used to represent the parity of atomic states etc.

AtomicLevels.Parity — Type.

struct Parity

Represents the parity of a quantum system, taking two possible values: even or odd.

The integer values that correspond to even and odd parity are +1 and -1, respectively. Base.convert can be used to convert integers into Parity values.

julia> convert(Parity, 1)
even

julia> convert(Parity, -1)
odd

AtomicLevels.@p_str — Macro.

@p_str

A string macro to easily construct Parity values.

julia> p"even"
even

julia> p"odd"
odd

The parity values also define an algebra and an ordering:

julia> p"odd" < p"even"
true

julia> p"even" * p"odd"
odd

julia> (p"odd")^3
odd

julia> -p"odd"
even

The exported parity function is overloaded for many of the types in AtomicLevels, defining a uniform API to determine the parity of an object.

AtomicLevels.parity — Function.

parity(object) -> Parity

Returns the parity of object.