DocTests

Documenter.DocTests — Module.

Provides the doctest function that makes sure that the jldoctest code blocks in the documents and docstrings run and are up to date.

Documenter.DocTests.doctest — Method.

doctest(blueprint, doc)

Traverses the pages and modules in the documenter blueprint, searching and executing doctests.

Will abort the document generation when an error is thrown. Use doctest = false keyword in Documenter.makedocs to disable doctesting.