DocChecks

Documenter.DocChecks — Module.

Provides the missingdocs, footnotes and linkcheck functions for checking docs.

Documenter.DocChecks.footnotes — Method.

footnotes(doc)

Checks footnote links in a Documents.Document.

Documenter.DocChecks.linkcheck — Method.

linkcheck(doc)

Checks external links using curl.

Documenter.DocChecks.missingdocs — Method.

missingdocs(doc)

Checks that a Documents.Document contains all available docstrings that are defined in the modules keyword passed to Documenter.makedocs.

Prints out the name of each object that has not had its docs spliced into the document.